Comments and saving csv

+++ ark/phenotyping/preprocess.py
@@ -52,9 +52,11 @@ def create_pixel_matrix(img_xr, seg_labels, fovs=None, channels=None, blur_facto
         # subset img_xr with only the fov we're looking for
         img_data_blur = img_xr.loc[fov, ..., channels].values
 
+        # for each marker, compute the Gaussian blur and cast to np.float32
         for marker in range(len(channels)):
             img_data_blur[:, :, marker] = ndimage.gaussian_filter(img_data_blur[:, :, marker],
-                                                                  sigma=blur_factor)
+                                                                  sigma=blur_factor).astype(
+                                                                  np.float32)
 
         # flatten each image
         pixel_mat = img_data_blur.reshape(-1, len(channels))
